requirements.txt
streamlit>=1.28.0
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.15
sentence-transformers>=2.2.2
google-search-results>=2.4.2
python-dotenv>=1.0.0
pydantic>=2.5.0
tiktoken>=0.5.2
pinecone>=3.0.0
pypdf2>=3.0.0
pdfplumber>=0.10.0
unstructured>=0.11.0
unstructured[pdf]>=0.11.0
pillow>=10.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
markdownify>=0.11.6
pyyaml>=6.0.0

